+++ locales/en/common/footer.js
@@ -1,7 +1,7 @@
 var footer =
     {
-    "copyright" : "Copyright © MapleFlagMissions.ca - All rights reserved<br><a href=\"/privacy.txt\" target=\"_blank\">Privacy Policy</a> - <a href=\"/agreement.txt\" target=\"_blank\">User Agreement</a>",
-    "update"    : "Last revised: 2nd December 2015",
+    "content" : "Copyright © MapleFlagMissions.ca - All rights reserved<br><a href=\"/privacy.txt\" target=\"_blank\">Privacy Policy</a> - <a href=\"/agreement.txt\" target=\"_blank\">User Agreement</a><br><a href=\"/checksums.txt\" target=\"_blank\">Checksums</a>",
+    "update"    : "Last revised: 7th December 2015",
     "credit"    : "Coded by <a href=\"https://github.com/Gliptal\" target=\"_blank\">Mattia Affabris</a> with <a href=\"http://www.sublimetext.com/3\" target=\"_blank\">Sublime Text 3</a>"
     }