+++ locales/en/help.js
@@ -1,10 +1,10 @@
 var help =
     {
-    "title"  : "Maple Flag HELP",
-    "help":
+    "title" : "Maple Flag HELP",
+    "help"  :
         {
-        "intro": "DCS can be very intimidating to new users. Many of the modules are considered \"study\" sims that require in depth knowledge and training to be able to make full use of them. Our Qualification Campaigns are a set of missions to test your knowledge and techniques to operate the various aircraft or vehicles. They should be attempted after you have some knowledge and understanding of the aircraft or vehicle systems.",
-        "list" :
+        "intro" : "DCS can be very intimidating to new users. Many of the modules are considered \"study\" sims that require in depth knowledge and training to be able to make full use of them. Our Qualification Campaigns are a set of missions to test your knowledge and techniques to operate the various aircraft or vehicles. They should be attempted after you have some knowledge and understanding of the aircraft or vehicle systems.",
+        "list"  :
             [
             "Read the DCS aircraft or vehicle manuals and review our Amplified Checklists provided with some of the Qualification Campaigns.",
             "Watch the training videos and watch other videos that others have done.",